# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Elden Ring API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/cook_books": {
            "get": {
                "description": "List Cook Books Alphabetically",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cook Books"
                ],
                "summary": "Get Cook Books",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CookBook"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a Cook Book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cook Books"
                ],
                "summary": "Create a Cook Book",
                "parameters": [
                    {
                        "description": "New Cook Book Params",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NameDescription"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CookBook"
                        }
                    }
                }
            }
        },
        "/api/cook_books/{CookBookID}": {
            "get": {
                "description": "Finds a Cook Book by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cook Books"
                ],
                "summary": "Get a Cook Book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cook Book ID",
                        "name": "CookBookID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CookBook"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Cook Book by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cook Books"
                ],
                "summary": "Delete a Cook Book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cook Book ID",
                        "name": "CookBookID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "description": "Update a Cook Book by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cook Books"
                ],
                "summary": "Update a Cook Book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cook Book ID",
                        "name": "CookBookID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Cook Book Params",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NameDescription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CookBook"
                        }
                    }
                }
            }
        },
        "/api/cook_books/{CookBookID}/recipe/{RecipeID}": {
            "post": {
                "description": "Adds a Recipe to Cook Book by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cook Books"
                ],
                "summary": "Create a Cook Book Recipe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cook Book ID",
                        "name": "CookBookID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "RecipeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CookBookRecipe"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Recipe from Cook Book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cook Books"
                ],
                "summary": "Delete a Cook Book Recipe",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cook Book ID",
                        "name": "CookBookID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "RecipeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/materials": {
            "get": {
                "description": "List Materials Alphabetically",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Get Materials",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Material"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a New Material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Create Material",
                "parameters": [
                    {
                        "description": "New Material Params",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NameDescription"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Material"
                        }
                    }
                }
            }
        },
        "/api/materials/{MaterialId}": {
            "get": {
                "description": "Get a Material by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Get a Material",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Material ID",
                        "name": "MaterialId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Material"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Material by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Delete a Material",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Material ID",
                        "name": "MaterialId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "description": "Update a Material by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Materials"
                ],
                "summary": "Update a Material",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Material ID",
                        "name": "MaterialId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Material Params",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NameDescription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Material"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.NameDescription": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.CookBook": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "recipes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Recipe"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.CookBookRecipe": {
            "type": "object",
            "properties": {
                "cookBookID": {
                    "type": "integer"
                },
                "recipe": {
                    "$ref": "#/definitions/models.Recipe"
                },
                "recipeID": {
                    "type": "integer"
                }
            }
        },
        "models.Material": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Recipe": {
            "type": "object",
            "properties": {
                "cookBooks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CookBook"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "materials": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RecipeMaterial"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.RecipeMaterial": {
            "type": "object",
            "properties": {
                "material": {
                    "$ref": "#/definitions/models.Material"
                },
                "materialID": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "recipeID": {
                    "type": "integer"
                }
            }
        }
    }
}